Add TaskRepository interface

+++ app/src/main/kotlin/de/robinwersich/todue/data/database/TaskDao.kt
@@ -9,6 +9,6 @@ interface TaskDao {
   @Insert suspend fun insert(task: Task)
   @Update suspend fun update(task: Task)
   @Delete suspend fun delete(task: Task)
-  @Query("SELECT * FROM todo WHERE id = :id") fun getTask(id: Long): Flow<Task>
+  @Query("SELECT * FROM todo WHERE id = :id") fun getTask(id: Int): Flow<Task>
   @Query("SELECT * FROM todo") fun getAllTasks(): Flow<List<Task>>
 }
